refactor: do major notification settings refactoring
